+++ R/get_density.R
@@ -20,6 +20,13 @@
 #' @export
 get_density <- function(result, timestep = result$max_i) {
 
+  # check if mdl_rn is provided
+  if (!inherits(x = result, what = "mdl_rn")) {
+
+    stop("Please prove mdl_rn object createt with run_simulation.", call. = FALSE)
+
+  }
+
   i <- timestep
 
   # check if i can be divided by save_each without reminder