+++ NEWS.md
@@ -3,6 +3,7 @@
 * Add `extract_result` function and `extract` argument to `run_simulation`
 * Add `bg_gamma` and `ag_gamma` to parameters. Was hard-coded previously
 * Better checking if object is `mdl_rn`
+* Renamed master branch to main
 
 # arrR 0.2.7
 * Add default parameters and starting values as data